% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_scfa.R
\name{default_scfa_analytes}
\alias{default_scfa_analytes}
\title{Default SCFA analyte names used in the Microbiome Insights workflow}
\usage{
default_scfa_analytes()
}
\value{
Character vector with ordered analyte names.
}
\description{
Default SCFA analyte names used in the Microbiome Insights workflow
}
